refactor: remove temp in file name
